+++ aldrin/src/handle.rs
@@ -38,7 +38,6 @@ use request::{
 };
 use std::future::Future;
 use std::hash::Hash;
-use std::mem::MaybeUninit;
 use std::num::NonZeroU32;
 use std::pin::Pin;
 use std::task::{Context, Poll};
@@ -680,32 +679,10 @@ impl Handle {
             .build_current_only()
             .await?;
 
-        let Some(event) = discoverer.next_event().await else {
-            return Ok(None);
-        };
-
-        if let Some(object) = object {
-            debug_assert_eq!(event.object_id().uuid, object);
-        }
-
-        // SAFETY: This creates an array of MaybeUninit, which doesn't require initialization.
-        let mut ids: [MaybeUninit<ServiceId>; N] = unsafe { MaybeUninit::uninit().assume_init() };
-
-        for (&uuid, id) in services.iter().zip(&mut ids) {
-            id.write(event.service_id(&discoverer, uuid));
-        }
-
-        // SAFETY: All N elements have been initialized in the loop above.
-        //
-        // In some future version of Rust, all this can be simplified; see:
-        // https://github.com/rust-lang/rust/issues/96097
-        // https://github.com/rust-lang/rust/issues/61956
-        let ids = unsafe {
-            (*(&MaybeUninit::new(ids) as *const _ as *const MaybeUninit<[ServiceId; N]>))
-                .assume_init_read()
-        };
-
-        Ok(Some((event.object_id(), ids)))
+        Ok(discoverer
+            .next_event()
+            .await
+            .map(|ev| (ev.object_id(), ev.service_ids_n(&discoverer, services))))
     }
 
     /// Finds any object implementing a set of services.
@@ -746,32 +723,11 @@ impl Handle {
             .build()
             .await?;
 
-        let Some(event) = discoverer.next_event().await else {
-            return Err(Error::Shutdown);
-        };
-
-        if let Some(object) = object {
-            debug_assert_eq!(event.object_id().uuid, object);
-        }
-
-        // SAFETY: This creates an array of MaybeUninit, which doesn't require initialization.
-        let mut ids: [MaybeUninit<ServiceId>; N] = unsafe { MaybeUninit::uninit().assume_init() };
-
-        for (&uuid, id) in services.iter().zip(&mut ids) {
-            id.write(event.service_id(&discoverer, uuid));
-        }
-
-        // SAFETY: All N elements have been initialized in the loop above.
-        //
-        // In some future version of Rust, all this can be simplified; see:
-        // https://github.com/rust-lang/rust/issues/96097
-        // https://github.com/rust-lang/rust/issues/61956
-        let ids = unsafe {
-            (*(&MaybeUninit::new(ids) as *const _ as *const MaybeUninit<[ServiceId; N]>))
-                .assume_init_read()
-        };
-
-        Ok((event.object_id(), ids))
+        discoverer
+            .next_event()
+            .await
+            .map(|ev| (ev.object_id(), ev.service_ids_n(&discoverer, services)))
+            .ok_or(Error::Shutdown)
     }
 
     /// Wait for any object implementing a set of services.